Column bind (cbind) two GctMatrix objects

Column bind (cbind) two GctMatrix objects

cbindGct(
  gctMatrix1,
  gctMatrix2,
  feature = c("union", "intersection"),
  missingValue = 0
)

Arguments

gctMatrix1

The first object

gctMatrix2

The second object

feature

What happens if the set of the features in both objects differ? Either union or intersection is possible.

missingValue

Missing values, NA or numeric values (such as 0) are accepted

Value

A larger matrix, with gctMatrix1 on the left and gctMatrix2 on the right, with merged features and descriptions.

Examples

m1 <- matrix(1:6, nrow=3, dimnames=list(sprintf("G%d", 1:3), sprintf("S%d", 1:2)))
m2 <- matrix(c(9:7, 12:10), nrow=3, dimnames=list(sprintf("G%d", 3:1), sprintf("S%d", 3:4)))
gm1 <- GctMatrix(m1, desc=sprintf("Gene%d", 1:3))
gm2 <- GctMatrix(m2, desc=sprintf("Gene%d", 3:1))
gm1
#> A GctMatrix with 3 features and 2 samples.
#> -- Below are the first rows and columns, use `showAll=TRUE` to print all values
#>    S1 S2
#> G1  1  4
#> G2  2  5
#> G3  3  6
#> First feature descriptions: Gene1,Gene2,Gene3 
#> -- Use `gctDesc` to fetch feature descriptions
#> -- Use `rownames` and `colnames` to fetch row and column names
#> -- All methods applicable to matrix can be used
gm2
#> A GctMatrix with 3 features and 2 samples.
#> -- Below are the first rows and columns, use `showAll=TRUE` to print all values
#>    S3 S4
#> G3  9 12
#> G2  8 11
#> G1  7 10
#> First feature descriptions: Gene3,Gene2,Gene1 
#> -- Use `gctDesc` to fetch feature descriptions
#> -- Use `rownames` and `colnames` to fetch row and column names
#> -- All methods applicable to matrix can be used
gm12 <- cbindGct(gm1, gm2)
gm12
#> A GctMatrix with 3 features and 4 samples.
#> -- Below are the first rows and columns, use `showAll=TRUE` to print all values
#>    S1 S2 S3 S4
#> G1  1  4  7 10
#> G2  2  5  8 11
#> G3  3  6  9 12
#> First feature descriptions: Gene1,Gene2,Gene3 
#> -- Use `gctDesc` to fetch feature descriptions
#> -- Use `rownames` and `colnames` to fetch row and column names
#> -- All methods applicable to matrix can be used
m3 <- matrix(13:18, nrow=3, dimnames=list(sprintf("G%d", 2:4), sprintf("S%d", 5:6)))
gm3 <- GctMatrix(m3, desc=sprintf("Gene%d", 2:4))
gm3
#> A GctMatrix with 3 features and 2 samples.
#> -- Below are the first rows and columns, use `showAll=TRUE` to print all values
#>    S5 S6
#> G2 13 16
#> G3 14 17
#> G4 15 18
#> First feature descriptions: Gene2,Gene3,Gene4 
#> -- Use `gctDesc` to fetch feature descriptions
#> -- Use `rownames` and `colnames` to fetch row and column names
#> -- All methods applicable to matrix can be used
gm123Intersect <- cbindGct(gm12, gm3, feature="intersect")
print(gm123Intersect, showAll=TRUE)
#> A GctMatrix with 2 features and 6 samples.
#>    S1 S2 S3 S4 S5 S6
#> G2  2  5  8 11 13 16
#> G3  3  6  9 12 14 17
#> attr(,"desc")
#> [1] "Gene2" "Gene3"
#> attr(,"class")
#> [1] "GctMatrix" "matrix"   
#> -- Use `gctDesc` to fetch feature descriptions
#> -- Use `rownames` and `colnames` to fetch row and column names
#> -- All methods applicable to matrix can be used
gm123Union <- cbindGct(gm12, gm3, feature="union")
print(gm123Union, showAll=TRUE)
#> A GctMatrix with 4 features and 6 samples.
#>    S1 S2 S3 S4 S5 S6
#> G1  1  4  7 10  0  0
#> G2  2  5  8 11 13 16
#> G3  3  6  9 12 14 17
#> G4  0  0  0  0 15 18
#> attr(,"desc")
#> [1] "Gene1" "Gene2" "Gene3" "Gene4"
#> attr(,"class")
#> [1] "GctMatrix" "matrix"   
#> -- Use `gctDesc` to fetch feature descriptions
#> -- Use `rownames` and `colnames` to fetch row and column names
#> -- All methods applicable to matrix can be used
gm123UnionNA <- cbindGct(gm12, gm3, feature="union", missingValue = NA)
print(gm123UnionNA)
#> A GctMatrix with 4 features and 6 samples.
#> -- Below are the first rows and columns, use `showAll=TRUE` to print all values
#>    S1 S2 S3 S4 S5 ...
#> G1  1  4  7 10 NA ...
#> G2  2  5  8 11 13 ...
#> G3  3  6  9 12 14 ...
#> G4 NA NA NA NA 15 ...
#> First feature descriptions: Gene1,Gene2,Gene3,Gene4 
#> -- Use `gctDesc` to fetch feature descriptions
#> -- Use `rownames` and `colnames` to fetch row and column names
#> -- All methods applicable to matrix can be used