Get file names for data import/export

Get file names for data import/export

iofile(x = NULL)

Arguments

x

File or directory name

Quite often we need to import and export data (especially bulky files) into a directory other than the local file. This function is a shortcut to get full names of import/export files.

The function first determines whether the option dataDir in the options of ribiosIO exists. If yes, its value will be used as the directory from/to which input/export files should be read/written.

If the value does not exist yet, the function tries to use a folder named data in the current working directory as dataDir. If this local folder exists, its name will be assigned to the dataDir option. If the folder does not exist, the function will report an error and quit.

The steps above guarantees that there is an option named dataDir, pointing to a directory where files are read from or written to.

The parameter x can be file or directory names in the dataDir directory. In this case, iofile(x) returns their full names. When x is missing or NULL, iofile() returns the value of dataDir. A common usage for the later case is dir(iofile()).

Examples

setDataDir(system.file("extdata", package="ribiosIO"))
dir(iofile())
#>  [1] "ampliseq-data"                              
#>  [2] "biokit_expression_files"                    
#>  [3] "example-DAVID-output-subset.txt"            
#>  [4] "subsetGctByGmt_data"                        
#>  [5] "test.chip"                                  
#>  [6] "test.cls"                                   
#>  [7] "test.gct"                                   
#>  [8] "test.gmt"                                   
#>  [9] "testSampleInfo-cov.txt"                     
#> [10] "testSampleInfo.txt"                         
#> [11] "test_colnames.gct"                          
#> [12] "test_colnames.txt"                          
#> [13] "test_lessStrict.gct"                        
#> [14] "test_nonnumbers.gct"                        
#> [15] "test_nonnumbers.txt"                        
#> [16] "test_read_exprs_matrix.gct"                 
#> [17] "test_read_exprs_matrix.tsv"                 
#> [18] "test_read_exprs_matrix.txt"                 
#> [19] "test_read_exprs_matrix_desc.tsv"            
#> [20] "test_read_exprs_matrix_duprownames.tsv"     
#> [21] "test_read_exprs_matrix_duprownames.txt"     
#> [22] "test_read_exprs_matrix_full.tsv"            
#> [23] "test_read_exprs_matrix_full.txt"            
#> [24] "test_read_exprs_matrix_full_duprownames.tsv"
#> [25] "test_read_exprs_matrix_full_duprownames.txt"
readLines(iofile("test.gct"), n=2)
#> [1] "#1.2"    "500\t26"