Read bedcov output of AmpliSeq amplicons and convert them to read counts

Read bedcov output of AmpliSeq amplicons and convert them to read counts

read_ampliseq_bedcovgct(file, bedFile)

Arguments

file

Character string, a GCT file containing bedcov output of amplicons

bedFile

Character string, an annotated BED file encoding amplicons

Value

A GctMatrix object containing read counts

The function is used to convert read base counts returned by samtools bedcov to read counts using Amplicon information encoded in the bed file

See also

Examples

bedlines <- paste0("#track type=bedDetail ionVersion=4.0 name=\"IAD50039-4_IAD87652-4_Design\"",
 "solution_type=4 description=\"TargetRegions_AmpliSeqID_IAD50039 AmpliSeq_Version=3.0.1",
 " Workflow=RNA merged with TargetRegions_AmpliSeqID_IAD87652 AmpliSeq_Version=4.48 Workflow=RNA\"",
 " color=77,175,74 priority=2", "\n",
 "NM_000014\t3316\t3421\tAMPL1384\t.\tGENE_ID=A2M;EntrezGeneID=2", "\n",
 "NM_005502\t2488\t2589\tAMPL28385508\t.\tGENE_ID=ABCA1;EntrezGeneID=19","\n",
 "NM_000927\t2520\t2624\tAMPL5599607\t.\tGENE_ID=ABCB1;EntrezGeneID=5243","\n",
 "NM_000443\t1367\t1470\tAMPL5513474\t.\tGENE_ID=ABCB4;EntrezGeneID=5244")
 gctLines <- paste0("#1.2", "\n",
 "3\t3","\n",
 "NAME\tDescription\tS1\tS2\tS3","\n",
 "A2M\tNM_000014\t105\t210\t315", "\n",
 "ABCA1\tNM_005502\t202\t303\t404", "\n",
 "ABCB1\tNM_000927\t312\t416\t520")
 bedcovGct <- read_ampliseq_bedcovgct(textConnection(gctLines), 
   textConnection(bedlines))
 bedcovGct
#> A GctMatrix with 3 features and 3 samples.
#> -- Below are the first rows and columns, use `showAll=TRUE` to print all values
#>       S1 S2 S3
#> A2M    1  2  3
#> ABCA1  2  3  4
#> ABCB1  3  4  5
#> First feature descriptions: NM_000014,NM_005502,NM_000927 
#> -- Use `gctDesc` to fetch feature descriptions
#> -- Use `rownames` and `colnames` to fetch row and column names
#> -- All methods applicable to matrix can be used