qRead BioKit expression file into a data.frame

qRead BioKit expression file into a data.frame

read_biokit_exprs(filename)

Arguments

filename

A BioKit expression file

The function uses an efficient C routine to read BioKit expression files. An Roche NGS expression file is eseentially a tab-delimited file. THe first six columns are mandatory (feature/tag name, multiple mapping RPKM, multiple mapping read count, unique mapping RPKM, unique mapping read count, and multiple mapping proportion). Right to these columns there can be arbitrary numbers of columns appended to annotate the features. In the current output, rows may have different numbers of columns: particularly for features without corresponding items in the annotation file used in the pipeline, their rows will contain the mandatory columns plus one extra column with the value “unknown”. This is handled automatically by the function.

Value

A data.frame contains both mandatory and additional columns. The first column of the expression file will be used as the row names of the data.frame object.

See also

Examples

biokitExampleFile <- system.file("extdata/biokit_expression_files/biokit-output-1.expression",
package="ribiosIO")
biokitExprs <- read_biokit_exprs(biokitExampleFile)