Read an expression matrix from file

Read an expression matrix from file. The file is either a GCT format file, a tab-delimited file or a space-delimited file.

read_exprs_matrix(x)

Arguments

x

File name

Value

A matrix

Details

An expression matrix of size m x n contains exprssion levels of m features in n samples. This function supports three commonly used file formats for expression levels: GCT format, tab-delimited file and space-delimited file.

Note

The function uses a very simple logic to guess whether the file is tab-delimited or space-delimited: it reads in the first n lines (currently n=3), and checks whether there is any tab character (\t): if yes, the file is parsed as tab-delmited, otherwise as space-delimited. Therefore, a space-delimited file should not contain tabs in case it needs to be parsed.

From ribiosIO version 1.0.2, this function supports duplicated row names.

From ribiosIO version 1.0-21, this function supports matrix file in which the second column is not numeric. This can happen, for instance, if the user decides to include descriptions. If such descriptions are detected, they are stored in the attribute “desc” so as to be later written into gct files.

From ribiosIO version 1.0-39, the function tolerates non-numeric values (such as '5?') in tab-delimited files better. However note that such values in the second column will cause problem because they will make the program interpret the second column as description but not numeric values.

See also

Author

Jitao David Zhang <jitao_david.zhang at roche.com>

Examples

testfile.path <- system.file("extdata", package="ribiosIO")

## import gct
read_exprs_matrix(file.path(testfile.path,"test_read_exprs_matrix.gct"))
#>   a b c
#> a 1 2 3
#> b 2 3 4
#> c 3 4 5
#> attr(,"desc")
#> [1] "" "" ""

## import tab-separated file
read_exprs_matrix(file.path(testfile.path, "test_read_exprs_matrix.tsv"))
#>   a b c
#> a 1 2 3
#> b 2 3 4
#> c 3 4 5

## import space-separated file
read_exprs_matrix(file.path(testfile.path, "test_read_exprs_matrix.txt"))
#>   a b c
#> a 1 2 3
#> b 2 3 4
#> c 3 4 5

## import tab-separated file with descriptions
read_exprs_matrix(file.path(testfile.path, "test_read_exprs_matrix_desc.tsv"))
#>   a b c
#> a 1 2 3
#> b 2 3 4
#> c 3 4 5
#> attr(,"desc")
#> [1] "A3" "B2" "T9"

## import tab-separated file with non-numeric values
read_exprs_matrix(file.path(testfile.path, "test_nonnumbers.txt"))
#>       Sample1 Sample2 Sample3
#> Gene1       1       2       3
#> Gene2       2       3       4
#> Gene3       4       6       1
#> Gene4       5       3       2
#> Gene5       5      NA       1
#> Gene6       1      NA       2
#> attr(,"desc")
#> [1] "G1" "G2" "G3" "G4" "G5" "G6"