Read pheno data from CLS or tab-delimited file

Read pheno (sample annotation) data from CLS file or tab-delimited file (sample information file).

read_pheno(file)

read_pheno_factor(file)

Arguments

file

A CLS file or tab-delimited file

Value

read_pheno returns a data.frame containing sample annotations. In case of CLS input file, the data.frame contains two columns: Array (indices of arrays) and Class (classes indexed in the GCT file). In case of tab-delimited file, the file will be parsed into the data.frame, assuming the file having column names but no row names.

Details

read_pheno returns a data.frame.

read_pheno_factor returns a factor, indicating sample groups. If the input file is a tab-delimited file, it filters out columns which are identical for all samples and columns which are unique for each sample. Consequently the remaining covariates are concatenated by the underscore character to form a factor. See examples below

References

For CLS and sample information file formats, see the GenePattern file formats documentation at https://software.broadinstitute.org/cancer/software/genepattern/file-formats-guide.

See also

Author

Jitao David Zhang <jitao_david.zhang@roche.com>

Examples

testClsFile <- system.file("extdata/test.cls", package="ribiosIO")
testPhenoFile <- system.file("extdata/testSampleInfo.txt",
package="ribiosIO")

(clsPheno <- read_pheno(testClsFile))
#>    Array   Class
#> 1      1 Control
#> 2      2    Case
#> 3      3 Control
#> 4      4    Case
#> 5      5    Case
#> 6      6 Control
#> 7      7    Case
#> 8      8    Case
#> 9      9    Case
#> 10    10 Control
#> 11    11    Case
#> 12    12 Control
#> 13    13    Case
#> 14    14    Case
#> 15    15    Case
#> 16    16 Control
#> 17    17    Case
#> 18    18 Control
#> 19    19    Case
#> 20    20    Case
#> 21    21 Control
#> 22    22 Control
#> 23    23 Control
#> 24    24 Control
#> 25    25    Case
#> 26    26    Case
(txtPheno <- read_pheno(testPhenoFile))
#>    Array   Class
#> 1      1 Control
#> 2      2    Case
#> 3      3 Control
#> 4      4    Case
#> 5      5    Case
#> 6      6 Control
#> 7      7    Case
#> 8      8    Case
#> 9      9    Case
#> 10    10 Control
#> 11    11    Case
#> 12    12 Control
#> 13    13    Case
#> 14    14    Case
#> 15    15    Case
#> 16    16 Control
#> 17    17    Case
#> 18    18 Control
#> 19    19    Case
#> 20    20    Case
#> 21    21 Control
#> 22    22 Control
#> 23    23 Control
#> 24    24 Control
#> 25    25    Case
#> 26    26    Case

## read_pheno_factor
(clsPhenoClass <- read_pheno_factor(testClsFile))
#>  [1] Control Case    Control Case    Case    Control Case    Case    Case   
#> [10] Control Case    Control Case    Case    Case    Control Case    Control
#> [19] Case    Case    Control Control Control Control Case    Case   
#> Levels: Case Control
(txtPhenoClass <- read_pheno_factor(testPhenoFile))
#>  [1] Control Case    Control Case    Case    Control Case    Case    Case   
#> [10] Control Case    Control Case    Case    Case    Control Case    Control
#> [19] Case    Case    Control Control Control Control Case    Case   
#> Levels: Control Case

testPhenoFileCov <- system.file("extdata/testSampleInfo-cov.txt",package="ribiosIO")
read_pheno_factor(testPhenoFileCov)
#>  [1] Control_W Case_W    Control_M Case_W    Case_M    Control_W Case_M   
#>  [8] Case_W    Case_M    Control_W Case_M    Control_W Case_M    Case_M   
#> [15] Case_W    Control_W Case_W    Control_M Case_M    Case_M    Control_M
#> [22] Control_M Control_M Control_M Case_M    Case_W   
#> Levels: Control_W Case_W Control_M Case_M