Write a factor in the CLS format

Write a factor in the CLS format

write_factor(fac, con = stdout(), offset = 0, sep = c("\t", " "))

write_cls(fac, con = stdout(), offset = 0, sep = c("\t", " "))

Arguments

fac

A factor

con

Connection to write to

offset

he integer representing the first level, default is set to 0, for some software it can be set to 1

sep

Separator used in the CLS format, can be '\t' (recommended) or ' ' (not to be used when space exists in levels)

Note

The original CLS format specifies that both tab or space can be used as separators. This makes it unable to represent factors with sapces in levels. In order to accomodate CLS format for these factors, we propose using tab as separators in CLS files when encoding factors in R. The default setting of read_factor and write_factor uses tab.

See also

Examples

set.seed(1887)
tempfac <- factor(sample(LETTERS, 30, replace=TRUE), levels=sample(LETTERS))
tempfile <- tempfile()
write_factor(tempfac, tempfile)
readLines(tempfile)
#> [1] "30\t26\t1"                                                                                               
#> [2] "#\tO\tI\tT\tN\tG\tH\tX\tQ\tA\tW\tR\tU\tV\tE\tD\tK\tM\tB\tF\tP\tZ\tY\tC\tJ\tS\tL"                         
#> [3] "5\t23\t25\t7\t18\t1\t13\t24\t9\t3\t14\t13\t2\t16\t22\t21\t12\t5\t13\t5\t1\t12\t17\t1\t7\t2\t3\t3\t12\t10"
stopifnot(identical(tempfac, read_factor(tempfile)))